as11: initially allocate 16kB header metadata filler space
Workaround issue where creating an AS-11 file not in a single write
pass resulted in slightly less than 8kB filler because of the header
metadata updates when completing the file.
This was not an issue for the DPP Metadata App. because it uses the
single write pass method to calculate an md5 whilst writing.